urdu-media-ai-utils/urdu-media-ocr-vision-utils/pdf_to_ocr_utils/requirements.txt
absl-py==1.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
certifi==2021.5.30
cffi==1.15.1
chardet==5.0.0
click==8.0.4
cryptography==40.0.2
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
doxapy==0.9.2
entrypoints==0.3
et-xmlfile==1.1.0
gast==0.6.0
grpcio==1.48.2
h5py==3.1.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.1
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.6.0
nltk==3.6.7
notebook==6.4.10
numpy==1.19.5
opencv-python==4.5.1.48
openpyxl==3.1.3
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.1
parso==0.7.0
path.py==8.1.2
pdf2image==1.17.0
pdfminer.six==20211012
pdfplumber==0.6.0
pexpect==4.9.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.17.1
prompt-toolkit==3.0.36
protobuf==3.19.6
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.14.0
pyparsing==3.1.2
PyPDF2==3.0.1
pyrsistent==0.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==25.1.2
regex==2023.8.8
Send2Trash==1.8.3
simplegeneric==0.8.1
six==1.16.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.6.0
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typing_extensions==4.1.1
Wand==0.6.13
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.0.3
xlrd==2.0.1
zipp==3.6.0
